+++ CHANGELOG.md
@@ -89,6 +89,7 @@
 - *(sequencer)* Deduplicate forced transactions in pool (#3807)
 - *(prover)* Empty-codehash check (#3913)
 - *(coordinator)* Preserve Amsterdam execution payload fields (#3949)
+- *(misc)* Make timer stop() await in-flight task execution (#4042)
 
 ### 🚜 Refactor
 
@@ -166,3 +167,11 @@
 - *(misc)* Upgrade teku libs to 26.9.0 (#3990)
 - *(misc)* Fix code licence header (#4009)
 - *(misc)* Improv JVM flakiness test and gradle configs (#4021)
+- *(coordinator, maru, linea-besu)* Update jvm dependencies: (#4029)
+- *(misc)* Maru use root gradle configs (#4031)
+- *(misc)* Update gradle to v9.8 (#4032)
+- *(misc)* Use junit over kotlin test (#4045)
+- *(misc)* Fix gradle build correctness bugs (#4044)
+- *(misc)* Gradle improve tasks (#4046)
+- *(misc)* Decrease vertx threadpool sizes for testing (#4047)
+- *(misc)* CI Merge sequencer plugin unit and acceptance jobs (#4048)